Reorganized
Reorganized the files, added a demo folder with an exe to run the game. Once running, wait for a few seconds (say about 10) till all of the materials are set to black. Moved the unity project into a sepreate folder if anyone is interested in how I made the game.